+++ src/SmtpServer/Protocol/RcptCommand.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using SmtpServer.ComponentModel;
@@ -22,9 +23,20 @@ namespace SmtpServer.Protocol
         /// Constructor.
         /// </summary>
         /// <param name="address">The address.</param>
-        public RcptCommand(IMailbox address) : base(Command)
+        public RcptCommand(IMailbox address)
+            : this(address, new Dictionary<string, string>())
+        {
+        }
+
+        /// <summary>
+        /// Constructor.
+        /// </summary>
+        /// <param name="address">The address.</param>
+        /// <param name="parameters">The list of recipient parameters.</param>
+        public RcptCommand(IMailbox address, IReadOnlyDictionary<string, string> parameters) : base(Command)
         {
             Address = address;
+            Parameters = parameters ?? new Dictionary<string, string>();
         }
 
         /// <summary>
@@ -40,10 +52,15 @@ namespace SmtpServer.Protocol
 
             using var container = new DisposableContainer<IMailboxFilter>(mailboxFilter);
 
-            switch (await container.Instance.CanDeliverToAsync(context, Address, context.Transaction.From, cancellationToken).ConfigureAwait(false))
+            var canDeliverTo = container.Instance is IParameterizedMailboxFilter parameterizedMailboxFilter
+                ? parameterizedMailboxFilter.CanDeliverToAsync(context, Address, context.Transaction.From, Parameters, cancellationToken)
+                : container.Instance.CanDeliverToAsync(context, Address, context.Transaction.From, cancellationToken);
+
+            switch (await canDeliverTo.ConfigureAwait(false))
             {
                 case true:
                     context.Transaction.To.Add(Address);
+                    context.Transaction.Recipients.Add(new SmtpMessageRecipient(Address, Parameters));
                     await context.Pipe.Output.WriteReplyAsync(SmtpResponse.Ok, cancellationToken).ConfigureAwait(false);
                     return true;
 
@@ -59,5 +76,10 @@ namespace SmtpServer.Protocol
         /// Gets the address that the mail is to.
         /// </summary>
         public IMailbox Address { get; }
+
+        /// <summary>
+        /// The list of recipient parameters.
+        /// </summary>
+        public IReadOnlyDictionary<string, string> Parameters { get; }
     }
 }